requirements.txt
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
db-dtypes==1.5.0
fonttools==4.61.1
google-api-core==2.28.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-cloud-bigquery==3.39.0
google-cloud-bigquery-storage==2.36.0
google-cloud-core==2.5.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.0
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pandas-gbq==0.33.0
pillow==12.1.0
proto-plus==1.27.0
protobuf==6.33.2
psutil==7.2.1
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydata-google-auth==1.9.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
tabulate==0.9.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
